Try reconnecting to database if queries fail
The DB connection is held in the backend object, so it is
long-running. Apparently some versions of MySQL will time out
a connection after some amount of idle time.

This solution is a little hacky because it tries reconnecting
on *all* errors from queries, even though some might not be DB
connection issues.